✨ refactor: core systems and UI update
