+++ shared/index.ts
@@ -1,2 +1,2 @@
-export * from "./types/Weather";
-export * from "./types/AreEngineBox";
+export * from "./types/weather";
+export * from "./types/areEngineBox";
